CAMEL-20447: camel-platform-http-main should not depend on jolokia but load it via the plugin. (#13248)
